# swagger-api.yml
openapi: 3.0.0
info:
  title: BlockTweet
  version: 1.0.0
  description: API pour communiquer avec le serveur. Une authentification est nécessaire , passer en Authorization Bearer votre token
...
paths:
...
  /api/users/auth:
    post:
      summary: Authentifier un utilisateur avec une signature
      description: Permet d'authentifier un utilisateur via une signature Ethereum.
      operationId: authenticate
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                  example: "0x1234567890abcdef"
                signature:
                  type: string
                  example: "0x..."
                message:
                  type: string
                  example: "Login to MyApp"
      responses:
        '200':
          description: Utilisateur authentifié avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: "jwt_token"
        '401':
          description: Signature invalide
        '500':
          description: Erreur lors de l'authentification
...
  /api/users/profile/{address}:
    get:
      summary: Récupérer le profil d'un utilisateur
      description: Permet de récupérer les informations d'un utilisateur à partir de son adresse Ethereum.
      operationId: getProfile
      tags:
        - Users
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
            example: "0x1234567890abcdef"
      responses:
        '200':
          description: Profil récupéré avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Utilisateur non trouvé
        '500':
          description: Erreur lors de la récupération du profil
  /api/users/fromToken:
    get:
      summary: Récupérer l'utilisateur à partir du token JWT
      description: Permet de récupérer les informations de l'utilisateur en utilisant un token JWT.
      operationId: getFromToken
      tags:
        - Users
      responses:
        '200':
          description: Utilisateur récupéré avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Token invalide
        '500':
          description: Erreur lors de la récupération de l'utilisateur
  /api/users/update:
    put:
      summary: Mettre à jour le profil de l'utilisateur
      description: Permet de mettre à jour le profil de l'utilisateur (username, avatar, bio).
      operationId: updateProfile
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "john_doe_updated"
                avatar:
                  type: string
                  example: "https://example.com/new_avatar.jpg"
                bio:
                  type: string
                  example: "Updated bio"
      responses:
        '200':
          description: Profil mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Nom d'utilisateur déjà existant
        '404':
          description: Utilisateur non trouvé
        '500':
          description: Erreur lors de la mise à jour du profil
  /api/users/{address}/follow:
    post:
      summary: Suivre un utilisateur
      description: Permet à un utilisateur de suivre un autre utilisateur par adresse Ethereum.
      operationId: followUser
      tags:
        - Users
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
            example: "0x1234567890abcdef"
      responses:
        '200':
          description: Suivi réussi
        '400':
          description: Impossible de se suivre soi-même
        '404':
          description: Utilisateur à suivre non trouvé
        '500':
          description: Erreur interne du serveur
  /api/users/{address}/unfollow:
    delete:
      summary: Ne plus suivre un utilisateur
      description: Permet à un utilisateur de ne plus suivre un autre utilisateur.
      operationId: unfollowUser
      tags:
        - Users
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
            example: "0x1234567890abcdef"
      responses:
        '200':
          description: Ne plus suivre l'utilisateur avec succès
        '400':
          description: Impossible de se désabonner soi-même
        '404':
          description: Utilisateur à ne plus suivre non trouvé
        '500':
          description: Erreur interne du serveur
  /api/users/{address}/followers:
    get:
      summary: Obtenir les followers d'un utilisateur
      description: Permet de récupérer la liste des followers d'un utilisateur.
      operationId: getFollowers
      tags:
        - Users
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
            example: "0x1234567890abcdef"
      responses:
        '200':
          description: Liste des followers récupérée avec succès
        '404':
          description: Utilisateur non trouvé
        '500':
          description: Erreur interne du serveur
  /api/users/{address}/following:
    get:
      summary: Obtenir les utilisateurs suivis
      description: Permet de récupérer la liste des utilisateurs suivis par un utilisateur.
      operationId: getFollowing
      tags:
        - Users
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
            example: "0x1234567890abcdef"
      responses:
        '200':
          description: Liste des utilisateurs suivis récupérée avec succès
        '404':
          description: Utilisateur non trouvé
        '500':
          description: Erreur interne du serveur
  /api/users/{address}/dofollow:
    get:
      summary: Vérifier si l'utilisateur suit un autre utilisateur
      description: Permet de vérifier si l'utilisateur connecté suit un autre utilisateur par adresse.
      operationId: doFollow
      tags:
        - Users
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
            example: "0x1234567890abcdef"
      responses:
        '200':
          description: Résultat de la vérification de suivi
          content:
            application/json:
              schema:
                type: object
                properties:
                  isFollowing:
                    type: boolean
                    example: true
        '404':
          description: Utilisateur non trouvé
        '500':
          description: Erreur interne du serveur
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        address:
          type: string
          example: "0x1234567890abcdef"
        username:
          type: string
          example: "john_doe"
        avatar:
          type: string
          example: "https://example.com/avatar.jpg"
        bio:
          type: string
          example: "Blockchain enthusiast"
        interests:
          type: array
          items:
            $ref: '#/components/schemas/Interest'
    Interest:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "blockchain"